Now mostly fully parses OK(HELLO) and also HELLO

+++ zerotier-network-hypervisor/src/lib.rs
@@ -2,7 +2,7 @@
 
 pub const VERSION_MAJOR: u8 = 1;
 pub const VERSION_MINOR: u8 = 99;
-pub const VERSION_REVISION: u8 = 1;
+pub const VERSION_REVISION: u16 = 1;
 
 pub mod error;
 pub mod util;
